requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-generativeai
python-dotenv
python-multipart
pdfplumber
fuzzywuzzy
python-Levenshtein
python-calamine
